+++ src/Main.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from '../src/components/App';
-import FolderPage from './components/Pages/FolderPage';
+import FolderPage from './components/Pages/FolderPage/FolderPage';
 import SharedPage from './components/Pages/SharedPage/SharedPage';
 // import HomePage from "./components/HomePage";
