+++ src/components/Pages/FolderPage/FolderPage.js
@@ -1,11 +1,10 @@
-import { useState, useEffect } from 'react';
-import { getFolder, getLink } from '../../api';
+import { useState, useEffect } from "react";
+import { getFolder, getLink } from "../../../api";
 import './FolderPage.css';
-import CardList from '../CardList';
-import linkImg from '../../assets/link-Img.png';
-import SearchBar from "../SearchBar";
-import FolderList from "../FolderList";
-
+import CardList from "../../CardList";
+import FolderList from "../../FolderList";
+import SearchBar from "../../SearchBar";
+import linkImg from "../../../assets/link-Img.png";
 
 function FolderPage() {
   const [folderList, setFolderList] = useState([]);
